_DataIssuesBuilder into the helper_factory.py
